requirements.txt
matplotlib
Pillow
numpy
scikit-learn
pyperclip